pyproject.toml
[project]
name = "actions"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.0",
    "requests>=2.32.5",
]
